packfile: use object_id in find_pack_entry_one()
The main function we use to search a pack index for an object is
find_pack_entry_one(). That function still takes a bare pointer to the
hash, despite the fact that its underlying bsearch_pack() function needs
an object_id struct. And so we end up making an extra copy of the hash
into the struct just to do a lookup.

As it turns out, all callers but one already have such an object_id. So
we can just take a pointer to that struct and use it directly. This
avoids the extra copy and provides a more type-safe interface.

The one exception is the get_delta_base() in packfile.c, when we are
chasing a REF_DELTA from inside the pack (and thus we have a pointer
directly to the mmap'd pack memory, not a struct). We can just bump the
hashcpy() from inside find_pack_entry_one() to this one caller that
needs it.