t: adjust root refs that violate syntax rules
The syntax of root refs (outside of "refs/") generally requires that
names end in "_HEAD" (both in the glossary, and also in code via the
is_root_ref_syntax() function). We don't yet enforce that consistently,
and there are several tests that use names that violate the rules.

In preparation for tighter enforcement, let's switch these to
syntactically valid names. In many of these cases we probably could just
move the refs into refs/heads/ or elsewhere. But without investigating
each test, that runs the risk of violating what the test is trying to do
(e.g., like the case in t9300 that we adjusted in an earlier commit). So
let's just adjust the names, which ensures we aren't ruining any tests,
but future-proofs us.