t1401: avoid pointless use of root ref
We check that symbolic-ref fails when trying to delete a symref that
doesn't exist. We use the name "FOO", but that does not conform to our
root-ref syntax. But there's no need for this to be a root ref. We can
use a name in refs/heads/, which future-proofs us against increased
enforcement of the syntax.

And while we're at it, let's use a name which more obviously explains
its purpose ("does-not-exist").