t9300: follow root ref syntax for TEMP_TAG
We fast-import a ref named TEMP_TAG at the root of the refs namespace
(outside of "refs/"), but this violates our usual root-ref syntax.

It is intentional that this is done at the root level. The test comes
from ea08a6fd19 (Actually allow TAG_FIXUP branches in fast-import,
2007-08-02), which says:

    New test cases have been added to make sure we still detect very
    wrong branch names (e.g. containing [ or starting with .) and yet
    still permit reasonable names (e.g. TAG_FIXUP).

Our definition of "reasonable" has tightened over the years, and
TAG_FIXUP (or TEMP_TAG) do not match our allowed root-ref syntax (though
we do not yet enforce it consistently). It must be TAG_FIXUP_HEAD. So
let's use that, which I think matches the spirit of the test.

While we're here, let's also swap out "rm -f" for "update-ref -d" in the
cleanup code. The former was a noop when using the reftable backend. The
leftover refs don't seem to affect downstream tests, but this matches
the original intent.