t3200: fully qualify symref destination
We create a dangling symref that points to "nowhere", which is not by
itself a valid refname (it is not in the refs/ hierarchy, and it does
not meet the all-caps syntactic requirement for a root ref).

We don't enforce these rules yet, but let's update the test in
preparation.