symbolic-ref: tighten syntax check of target refname
When creating or updating a symref, we pass the destination refname to
check_refname_format() to make sure it is a valid refname. But because
we pass the ALLOW_ONELEVEL flag, that function can't enforce the rules
for root refs (it doesn't know if we are passing a fully qualified ref
and are OK with it being a top-level ref, or if we are passing a partial
component that resides within the refs/ namespace).

We know that symrefs always store fully qualified refnames. So we can
pass the new FULLY_QUALIFIED flag to let check_refname_format() know
that it can enforce the root-ref syntax rules.

The test shows off the change in behavior: we'll now forbid pointing to
a root ref named FOO (it would have to be FOO_HEAD or similar).