t1400: fix broken target of symref update
While testing "update-ref --stdin --batch-updates", we invoke it with
this input:

  update refs/heads/ref1 <some-oid> <current-oid>
  symref-update refs/heads/ref2 <some-oid> ref refs/heads/nonexistent

With the idea being that both ref1 and ref2 are regular, non-symref
refs, and we should refuse to update ref2 because it doesn't match the
expected "ref refs/heads/nonexistent" value.

But there's an oddity here: we are asking to do a symref-update to an
oid, not to a valid symref value. This is nonsense, and would point to a
bogus symref (a refname with the literal value of the oid). But because
we don't yet enforce ref syntax rules for symref targets, the test is
able to continue and rejects the ref for the reason it expects.

However, once a future patch starts enforcing those syntax rules, we'll
start rejecting the update of "ref2" for that reason instead.

Let's switch out the oid for a plausible ref value. The point of the
test is that we should never get as far as actually updating ref2 to
that value. By giving it a syntactically valid target, we future proof
against tighter enforcement of the rules.