update-ref: verify stdin refnames as fully qualified
When parsing refnames on stdin, we try to make sure they are
syntactically valid before handing them off to the ref transaction. But
we only check with the ALLOW_ONELEVEL flag, but they really should be
fully-qualified refs.

The ref code will eventually complain about these names, but if we are
going to bother checking them up front, we should do so using the
FULLY_QUALIFIED flag, which catches names like "invalid-at-root" which
are valid "onelevel" names (like "refs/heads/invalid-at-root") but not
actual root names.