strbuf: more explicit NUL termination
I suspect coverity thinks strings are not NUL-terminated anywhere that
we call strbuf_setlen(), because of its code-path to avoid writing into
the slopbuf. Even though there's an assert there, that doesn't seem to
be enough (though maybe a BUG() would do it?).

Anyway, it should be safe to manually terminate after each call as long
as we are in a function that definitely called strbuf_grow(). I hit a
new case here based on strbuf_splice(), but I'm hoping this will cleanup
a lot of extra false positives, too.