merged: show partially-merged topics
The current code just checks if there are any patches that aren't
present in the upstream branch, and counts that as "merged". But this
isn't always accurate, as sometimes applying the patch on a different
base shifts the patch-id enough to create a "new" patch.

Let's try to give a bit more information here by reporting "partial"
merges. These _could_ be topics that have extra patches that are
missing, but usually it's going to be a shifted patch-id. Likewise, a
single-patch topic with a shifted patch-id will still appear as
unmerged, even though it is merged.

This is all mostly informational, so it's OK. Topics are not generally
deleted based on this flag. Instead, as we rebase them forward, they
eventually turn into the empty set, and then are by definition "merged
to master".