merged: cache topic patch-ids
Just like we cached the upstream patch-id computation in the last
commit, we can do the same for the actual topics. We don't always use
these multiple times (if we find them merged we stop looking), but in
the worst case we may re-compute once for each upstream.

This shows an obvious improvement:

  Benchmark 1: Meta/merged-old
    Time (mean ± σ):     984.8 ms ±  82.8 ms    [User: 905.6 ms, System: 438.5 ms]
    Range (min … max):   863.9 ms … 1150.8 ms    10 runs

  Benchmark 2: Meta/merged
    Time (mean ± σ):     528.4 ms ±  21.2 ms    [User: 495.9 ms, System: 215.9 ms]
    Range (min … max):   502.0 ms … 556.6 ms    10 runs

  Summary
    Meta/merged ran
      1.86 ± 0.17 times faster than Meta/merged-old

But the difference gets even larger as you add more possible upstreams:

  Benchmark 1: Meta/merged-old --jch --seen
    Time (mean ± σ):      2.047 s ±  0.134 s    [User: 1.889 s, System: 0.851 s]
    Range (min … max):    1.813 s …  2.282 s    10 runs

  Benchmark 2: Meta/merged --jch --seen
    Time (mean ± σ):     921.7 ms ±  47.5 ms    [User: 796.2 ms, System: 359.0 ms]
    Range (min … max):   856.7 ms … 1001.6 ms    10 runs

  Summary
    Meta/merged --jch --seen ran
      2.22 ± 0.19 times faster than Meta/merged-old --jch --seen