merged: collect all topic patch-ids in a single process
This is a complete rewrite of the merged script. It switches to perl,
which is necessary to work with actual hashes. But the big change here
is that instead of running a separate traversal for each topic, we're
going to run a single one with "log --source".

Unfortunately there isn't a good way to feed "log --source" through
"patch-id", as the latter throws away the extra information on each
line. But we can at least do it all in two runs: one to map commits to
sources (where inverting gives us the commits in each topic) and one to
get the patch-id for each commit.

With those two, we know the patch-ids in each topic. And then grabbing
the patch-ids for each upstream let's us do an easy comparison. Here are
the results for 40 topics:

  Benchmark 1: Meta/merged-old
    Time (mean ± σ):     559.3 ms ±  43.1 ms    [User: 477.9 ms, System: 198.4 ms]
    Range (min … max):   507.5 ms … 662.8 ms    10 runs

  Benchmark 2: Meta/merged
    Time (mean ± σ):     240.2 ms ±   8.9 ms    [User: 198.7 ms, System: 50.6 ms]
    Range (min … max):   234.4 ms … 266.3 ms    11 runs

  Summary
    Meta/merged ran
      2.33 ± 0.20 times faster than Meta/merged-old

And adding in extra upstreams is similarly faster:

  Benchmark 1: Meta/merged-old --jch --seen
    Time (mean ± σ):     909.7 ms ±  53.6 ms    [User: 739.9 ms, System: 312.6 ms]
    Range (min … max):   852.1 ms … 1007.2 ms    10 runs

  Benchmark 2: Meta/merged --jch --seen
    Time (mean ± σ):     487.0 ms ±  18.0 ms    [User: 400.0 ms, System: 106.5 ms]
    Range (min … max):   477.9 ms … 537.6 ms    10 runs

  Summary
    Meta/merged --jch --seen ran
      1.87 ± 0.13 times faster than Meta/merged-old --jch --seen

I do minorly worry that --source is going to misattribute commits when
there are some shared between topics (e.g., topic A builds off of topic
B). I think we might be OK, though, because in such cases one topic is
typically a superset of another. And so we'd decorate the tips of each
topic individually. E.g., something like:

  git init
  for i in $(seq 10); do git commit -m "main $i" --allow-empty; done
  git checkout -b tip
  for i in $(seq 10); do git commit -m "tip $i" --allow-empty; done
  git log --source --oneline main tip

gets this right (all of "main" is labeled as such, even though we can
traverse there from "tip").

I'd worry more about the case where we come into a topic part-way, like:

  git init
  for i in $(seq 10); do git commit -m "main $i" --allow-empty; done
  git checkout -b side HEAD~5
  for i in $(seq 10); do git commit -m "side $i" --allow-empty; done
  git log --source --oneline main side

We _seem_ to get this right in practice, but I suspect it may be
sensitive to variations in traversal order (if we paint down to "main 5"
from "side 1" first rather than "main 6"). But I also doubt I'd ever do
that with my series (I'd break a topic into a base topic before I
branched off some middle commit).