merged: show "jch" by default
Now that the script is quite fast, it's probably reasonable to be more
thorough in the default outupt. So let's show things merged to "jch" by
default (which just lets us know that Junio has seen it and picked up
the series).

We could add "--seen", too, but I really don't think it's all that
interesting. And while the script is now pretty fast, it's not free
(it's probably 80ms to check each one).