merged: avoid shell in git-log invocation
Passing a list rather than a string lets us avoid any quoting issues
(which we do handle with quotemeta, but it's an ugly extra step). Some
of the command invocations used to require a shell (e.g., to pipe "log |
patch-id") but we switched to a single invocation in 7b9040c (merged:
compute patch-id and source in single git-log, 2025-08-27).

As a bonus, I think this also makes much more obvious the subtle pasting
issue discussed in that commit. Namely that we end up with:

  origin/$upstream...$topic1 $topic2

passed to the shell. Now we explicitly handle $topic1 separately.